add property filters

+++ database/migrations/2023_11_10_121019_create_property_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('judul');
             $table->string('status');
             $table->string('tipe_properti');
-            $table->integer('harga');
+            $table->unsignedBigInteger('harga');
             $table->text('deskripsi');
             $table->string('area');
             $table->foreignId('provinsi_id')->nullable()->constrained('provinsis');
